Replace tests with benchmarks
