+++ python/cusignal/conftest.py
@@ -19,6 +19,12 @@ import numpy as np
 # Fixtures with (scope="session") will execute once
 # and be shared will all tests that need it.
 
+def pytest_configure(config):
+    config.addinivalue_line(
+        "markers", "slow: mark CPU test cases"
+    )
+
+
 # Generate data for using linspace
 @pytest.fixture(scope="session")
 def linspace_data_gen():
