+++ python/setup.cfg
@@ -41,14 +41,6 @@ addopts =
 	--benchmark-warmup-iterations=10 
 	--benchmark-disable-gc
 	--benchmark-disable
+	--strict-markers
 markers =
-	slow:
-python_classes =
-	Bench*
-	Test*
-python_files =
-	bench_*
-	test_*
-python_functions = 
-	bench_*
-	test_*
+	slow: marks tests as slow (deselect with '-m "not slow"')
